requirements.txt
gTTS
SpeechRecognition
pygame
pytube
requests
python-vlc